+++ service_geo_location/src/it/resources/logback.xml
