--- service_musit_thing/test/resources/evolutions/default/1.sql
@@ -1,39 +1,0 @@
-#
-#  MUSIT is a museum database to archive natural and cultural history data.
-#  Copyright (C) 2016  MUSIT Norway, part of www.uio.no (University of Oslo)
-#
-#  This program is free software; you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation; either version 2 of the License,
-#  or any later version.
-#
-#  This program is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License along
-#  with this program; if not, write to the Free Software Foundation, Inc.,
-#  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-#
-# Example schema
- 
-# --- !Ups
-
-CREATE SCHEMA IF NOT EXISTS MUSIT_MAPPING;
-
--- TODO: remove for final musit thing
-CREATE TABLE MUSIT_MAPPING.VIEW_MUSITTHING (
-    ny_id bigint(20) NOT NULL AUTO_INCREMENT,
-    displayid varchar(255),
-    displayname varchar(255),
-    PRIMARY KEY (ny_id)
-);
-
-insert into MUSIT_MAPPING.VIEW_MUSITTHING (ny_id, displayid, displayname) values (NULL, 'C1', 'Øks5');
-insert into MUSIT_MAPPING.VIEW_MUSITTHING (ny_id, displayid, displayname) values (NULL, 'C2', 'Kniv7');
- 
-# --- !Downs
- 
-DROP TABLE VMUSIT_MAPPING.IEW_MUSITTHING;
-DROP SCHEMA MUSIT_MAPPING;
